--- app/src/main/kotlin/com/wire/android/feature/auth/client/datasource/DeviceType.kt
@@ -1,8 +1,0 @@
-package com.wire.android.feature.auth.client.datasource
-
-sealed class DeviceType
-
-object Permanent : DeviceType()
-object Temporary : DeviceType()
-object LegalHold : DeviceType()
-object Unknown : DeviceType()
