--- app/src/main/kotlin/com/wire/android/feature/auth/client/datasource/mapper/DeviceTypeMapper.kt
@@ -1,17 +1,0 @@
-package com.wire.android.feature.auth.client.datasource.mapper
-
-import com.wire.android.feature.auth.client.datasource.DeviceType
-import com.wire.android.feature.auth.client.datasource.LegalHold
-import com.wire.android.feature.auth.client.datasource.Permanent
-import com.wire.android.feature.auth.client.datasource.Temporary
-
-class DeviceTypeMapper {
-
-    fun toStringValue(type: DeviceType) : String =
-        when (type) {
-            Permanent -> "permanent"
-            Temporary -> "temporary"
-            LegalHold -> "legalhold"
-            else -> "unknown"
-        }
-}
