+++ app/src/test/kotlin/com/wire/android/feature/auth/client/datasource/mapper/DeviceTypeMapperTest.kt
@@ -10,7 +10,7 @@ import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Before
 import org.junit.Test
 
-class DeviceClassMapperTest : UnitTest() {
+class DeviceTypeMapperTest : UnitTest() {
 
     private lateinit var deviceTypeMapper: DeviceTypeMapper
